+++ include/RadonFramework/Core/Types/UInt32.hpp
@@ -7,7 +7,7 @@
 namespace RadonFramework { namespace Core { namespace Types {
 
 /// A UInt32 have always the size of 4 byte and represent a unsigned number.
-typedef unsigned long UInt32;
+typedef uint32_t UInt32;
 
 /// The minimum value which can be contained by a UInt32.
 extern const UInt32 UInt32Min;
